requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.2.post1
async-timeout==5.0.1
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.5
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pika==1.3.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
pymongo==4.10.1
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.2.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
